Extract Terraform commands
This commit extracts all terraform functions to a dedicated file. It adds parameters to function calls as necessary to allow current functionality to continue to work.
